requirements.txt
tensorflow
tensorflow_datasets
numpy
pandas
streamlit
scikeras